Revert "Rename dirs"
This reverts commit 5577962b17484cf887075d7e9783bf3b2ec15eb0.
