+++ lib/features/pill_sheet_modified_history/pill_sheet_modified_history_page.dart
@@ -5,8 +5,8 @@ import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/components/molecules/indicator.dart';
 import 'package:pilll/provider/database.dart';
-import 'package:pilll/features/calendar/components/pill_sheet_modified_history/components/pill_sheet_modified_history/pill_sheet_modified_history_list.dart';
-import 'package:pilll/features/calendar/components/pill_sheet_modified_history/components/pill_sheet_modified_history/pill_sheet_modified_history_list_header.dart';
+import 'package:pilll/features/calendar/components/pill_sheet_modified_history/pill_sheet_modified_history_list.dart';
+import 'package:pilll/features/calendar/components/pill_sheet_modified_history/pill_sheet_modified_history_list_header.dart';
 import 'package:pilll/features/error/universal_error_page.dart';
 import 'package:pilll/provider/pill_sheet_modified_history.dart';
 import 'package:pilll/provider/user.dart';
